define the name of the vm second time
Could be because of my unique combination of vagrant/virtualbox
/virtualbox plugin that the VMs don't get created with their names as
given in the `config.vm.define`, so I had to explicitly name them using
`cfg.vm.provider` block.

Not too keen on getting this merged, it could be only an issue on my
end, but I did say I'll create a PR so here it is.